// src/types/activity.type.ts
export interface EffortCamal {
  elapsedTime: number;
  endIndex: number;
  id: bigint;
  movingTime: number;
  name: string;
  startIndex: number;
}

export interface GearCamal {
  convertedDistance: number;
  distance: number;
  id: string;
  name: string;
  primary: boolean;
  retired: boolean;
}

export interface LapCamal {
  averageCadence: number;
  averageHeartrate: number;
  averageSpeed: number;
  distance: number;
  elapsedTime: number;
  endIndex: number;
  id: bigint;
  lapIndex: number;
  maxHeartrate: number;
  maxSpeed: number;
  movingTime: number;
  name: string;
  paceZone: number;
  split: number;
  startIndex: number;
  totalElevationGain: number;
}

export type SplitMetricCamal = SplitCamal;

export type SplitStandardCamal = SplitCamal;

export interface StravaActivity {
  athlete: Athlete;
  average_cadence: number;
  average_heartrate: number;
  average_speed: number;
  best_efforts: Effort[];
  device_name: string;
  distance: number;
  elapsed_time: number;
  elev_high: number;
  elev_low: number;
  end_latlng: [number, number];
  gear: Gear;
  has_heartrate: boolean;
  id: bigint;
  laps: Lap[];
  map: ActivityMap;
  max_heartrate: number;
  max_speed: number;
  moving_time: number;
  name: string;
  splits_metric: SplitMetric[];
  splits_standard: SplitStandard[];
  start_date: string;
  start_date_local: string;
  start_latlng: [number, number];
  total_elevation_gain: number;
  type: string;
}

export interface StravaActivityCamal {
  athlete: AthleteCamal;
  averageCadence: number;
  averageHeartrate: number;
  averageSpeed: number;
  bestEfforts: EffortCamal[];
  deviceName: string;
  distance: number;
  elapsedTime: number;
  elevHigh: number;
  elevLow: number;
  endLatlng: [number, number];
  gear: GearCamal;
  hasHeartrate: boolean;
  id: bigint;
  laps: LapCamal[];
  map: ActivityMapCamal;
  maxHeartrate: number;
  maxSpeed: number;
  movingTime: number;
  name: string;
  splitsMetric: SplitMetricCamal[];
  splitsStandard: SplitStandardCamal[];
  startDate: string;
  startDateLocal: string;
  startLatlng: [number, number];
  totalElevationGain: number;
  type: string;
}

interface ActivityMap {
  id: string;
  polyline: string;
  summary_polyline: string;
}

interface ActivityMapCamal {
  id: string;
  polyline: string;
  summaryPolyline: string;
}

interface Athlete {
  id: bigint;
}

interface AthleteCamal {
  id: number;
}

interface Effort {
  elapsed_time: number;
  end_index: number;
  id: bigint;
  moving_time: number;
  name: string;
  start_index: number;
}

interface Gear {
  converted_distance: number;
  distance: number;
  id: string;
  name: string;
  primary: boolean;
  retired: boolean;
}

interface Lap {
  average_cadence: number;
  average_heartrate: number;
  average_speed: number;
  distance: number;
  elapsed_time: number;
  end_index: number;
  id: bigint;
  lap_index: number;
  max_heartrate: number;
  max_speed: number;
  moving_time: number;
  name: string;
  pace_zone: number;
  split: number;
  start_index: number;
  total_elevation_gain: number;
}

interface Split {
  average_grade_adjusted_speed: number;
  average_heartrate: number;
  average_speed: number;
  distance: number;
  elapsed_time: number;
  elevation_difference: number;
  moving_time: number;
  pace_zone: number;
  split: number;
}

interface SplitCamal {
  averageGradeAdjustedSpeed: number;
  averageHeartrate: number;
  averageSpeed: number;
  distance: number;
  elapsedTime: number;
  elevationDifference: number;
  movingTime: number;
  paceZone: number;
  split: number;
}

type SplitMetric = Split;

type SplitStandard = Split;